Add auto-reply "Hello!" in OnMessageReceived method
The code change adds a new line within the `OnMessageReceived` method in the `Mattermost.ConsoleTest` namespace. Specifically, after logging the received message to the console, the code now sends a "Hello!" message back to the channel from which the message was received using the `SendMessageAsync` method of the `Client` object.