Make AuthData nullable in User class and update tests
The `AuthData` property in the `User` class within the `Mattermost.Models.Users` namespace has been changed from a non-nullable string (`string`) to a nullable string (`string?`). This change allows the `AuthData` property to hold a `null` value.

Correspondingly, the test assertion in the `MattermostClientTests` class within the `Mattermost.Tests` namespace has been updated to remove the check that `AuthData` is not null. This reflects the change in the `User` class, acknowledging that `AuthData` can now be `null`.